feature(add-spring-boot): задание 5 https://online.t1-academy.ru/viewer/sessions/5807/tasks/3400
